Improve code style and formatting
Remove the `this` assignment of the function parameter (it's unnecessary, not entirely sure why you were doing that but let me know if there's a good reason for it) and change some alignment to make the code more easily readable.

Also removed the require(express) statement as frontend JS uses import statements rather than CommonJS require, so this is only needed for backend devs. You can add details on imports in here later if you want.